agent_ai/requirements-api.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pydantic
langgraph
openai
pymysql
openpyxl